third_party/blink/web_tests/external/wpt/navigation-api/ordering-and-transition/location-href-intercept_currententrychange-expected.txt

This is a testharness.js-based test.
[FAIL] event and promise ordering for location.href setter intercepted by intercept()
  assert_array_equals: expected property 3 to be "navigatesuccess" but got "promise microtask" (expected array ["navigate", "currententrychange", "handler run", "navigatesuccess", "promise microtask", "transition.finished fulfilled"] got ["navigate", "currententrychange", "handler run", "promise microtask", "navigatesuccess", "transition.finished fulfilled"])
Harness: the test ran to completion.
